added basic view ticket view update
